allow zipped help and example files
git-svn-id: https://svn.r-project.org/R/branches/R-0-64-patches@4091 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/base/R/help.R
@@ -6,7 +6,6 @@ function (topic, offline = FALSE, package = c(.packages(), .Autoloaded),
           lib.loc = .lib.loc, verbose = .Options$verbose,
           htmlhelp = .Options$htmlhelp)
 {
-    ## && !is.character(package))
     if (!missing(package))
         if (is.name(y <- substitute(package)))
             package <- as.character(y)
@@ -19,17 +18,15 @@ function (topic, offline = FALSE, package = c(.packages(), .Autoloaded),
         # for cmd/help ..
         if (!is.na(match(topic, c("+", "-", "*", "/", "^", "%%"))))
             topic <- "Arithmetic"
-        else if (!is.na(match(topic, c("<", ">", "<=", ">=",
-            "==", "!="))))
+        else if (!is.na(match(topic, c("<", ">", "<=", ">=", "==", "!="))))
             topic <- "Comparison"
         else if (!is.na(match(topic, c("[", "[[", "$"))))
             topic <- "Extract"
-        else if (!is.na(match(topic, c("&", "&&", "|", "||",
-            "!"))))
+        else if (!is.na(match(topic, c("&", "&&", "|", "||", "!"))))
             topic <- "Logic"
         else if (!is.na(match(topic, c("%*%"))))
-            topic <- "matmult"
-        topic <- gsub("\\[", "\\\\[", topic)
+            topic<- "matmult"
+        topic <- gsub("\\[","\\\\[", topic)
         INDICES <- system.file(pkg=package, lib=lib.loc)
         file <- index.search(topic, INDICES, "AnIndex")
         if (file == "") {
@@ -39,8 +36,8 @@ function (topic, offline = FALSE, package = c(.packages(), .Autoloaded),
         }
         if (length(file) && file != "") {
             if (verbose)
-                cat("\t\t\t\t\t\tHelp file name `", sub(".*/",
-                  "", file), ".Rd'\n", sep = "")
+                cat("\t\t\t\t\t\tHelp file name `", sub(".*/", "", file),
+                    ".Rd'\n", sep = "")
             if (!offline) {
                 if (!is.null(htmlhelp) && htmlhelp) {
                   ## replace the last occurence of /help/ in the
@@ -57,25 +54,38 @@ function (topic, offline = FALSE, package = c(.packages(), .Autoloaded),
                   cat("help() for", topic, " is shown in browser",
                     browser, "...\n")
                 }
-                else file.show(file)
+                else {
+                    ## experimental code
+                    zfile <- zip.file.extract(file, "Rhelp.zip")
+                    ## end of experimental code
+                    file.show(zfile, delete.file = (zfile!=file))
+                }
             }
             else {
-                FILE <- tempfile()
-                on.exit(unlink(FILE))
-                cat("\\documentclass[", .Options$papersize, "paper]{article}\n",
-                  file = FILE, sep = "")
-                file.append(FILE,
-                            file.path(R.home(), "doc", "manual", "Rd.sty"))
-                cat("\\InputIfFileExists{Rhelp.cfg}{}{}\n\\begin{document}\n",
-                  file = FILE, append = TRUE)
-                file.append(FILE, paste(sub("help/", "latex/",
-                  file), ".tex", sep = ""))
-                cat("\\end{document}\n", file = FILE, append = TRUE)
-                system(paste(file.path(R.home(), "bin", "help"),
-                             "PRINT", FILE, topic,
-                             .Options$latexcmd, .Options$dvipscmd)
-                       )
-                return()
+                ltxfile <- paste(sub("help/", "latex/", file), ".tex", sep = "")
+                ## experimental code
+                zfile <- zip.file.extract(ltxfile, "Rhelp.zip")
+                if(zfile != ltxfile) on.exit(unlink(ltxfile))
+                ## end of experimental code
+                if(file.exists(zfile)) {
+                    FILE <- tempfile()
+                    on.exit(unlink(FILE))
+                    cat("\\documentclass[", .Options$papersize, "paper]{article}\n",
+                        file = FILE, sep = "")
+                    file.append(FILE,
+                                file.path(R.home(), "doc", "manual", "Rd.sty"))
+                    cat("\\InputIfFileExists{Rhelp.cfg}{}{}\n\\begin{document}\n",
+                        file = FILE, append = TRUE)
+                    file.append(FILE, zfile)
+                    cat("\\end{document}\n", file = FILE, append = TRUE)
+                    system(paste(file.path(R.home(), "bin", "help"),
+                                 "PRINT", FILE, topic,
+                                 .Options$latexcmd, .Options$dvipscmd)
+                           )
+                    return(invisible())
+                }
+                else
+                    stop(paste("No offline documentation for", topic, "is available"))
             }
         }
         else stop(paste("No documentation for `", topic, "'",
